Add an alternative name parser
The parser copied from OpenSSL has two problems:
1. it doesn't parse correctly RDN such as CN=unified/voms, i.e. a value
   that includes a '/' character
2. it doesn't manage correctly O values longer than 64 characters

The 2nd problem could be addressed setting a larger limit with
ASN1_STRING_TABLE_add. Unfortunately calling this function crashes
systematically under OpenSSL 1.0.2.

The 1st problem would require amending the parser.

This commit adds an alternative parser that correctly handles the
1st problem. It is less powerful than the original, in that it doesn't
handle multi-valued RDNs and escaped characters. Let's see if this
suffices.

The 2nd problem still needs to be addressed. It could be made to work
on OpenSSL 1.1.x, but that's not enough, since it has to work on 1.0.2.
I'm afraid the only option is to not use X509_NAME's.