Avoid the call to get_root_groups
The call produces a list with the top group (i.e. the VO) followed
by the top-level VOMS groups, failing if only the top group exists.
However the list is used only to get the first group (i.e. the VO)
to seed the recursive import of groups in IAM.

Fix #14